Factor out analyzer.ts from the parser.ts
There are distinct steps of first transforming the raw strings of an
input log into structured format, and then computing and analysing
BpfState. Reflect that in the code organization by introducing
analyzer.ts
